Refactored; now controllers routes are registered in startup with fewer magic strings. now support xml

+++ GeekMDSuite.WebAPI/Presentation/Controllers/EntityDataController.cs
@@ -8,8 +8,7 @@ using Microsoft.AspNetCore.Mvc;
 
 namespace GeekMDSuite.WebAPI.Presentation.Controllers
 {
-    [Route("api/[controller]")]
-    [Produces("application/json")]
+    [Produces("application/json", "application/xml")]
     public abstract class EntityDataController<T> : Controller where T : class, IEntity<T>
     {
         private readonly IRepository<T> _repo;
@@ -23,7 +22,15 @@ namespace GeekMDSuite.WebAPI.Presentation.Controllers
         }
 
         [HttpGet]
+        [Route("")]
         public IActionResult Get()
+        {
+            return NotFound("Not an API endpoint. Please see the documentation.");
+        }
+        
+        [HttpGet]
+        [Route("all/")]
+        public IActionResult GetAll()
         {
             try
             {
@@ -31,12 +38,12 @@ namespace GeekMDSuite.WebAPI.Presentation.Controllers
             }
             catch (RepositoryElementNotFoundException)
             {
-                return NotFound();
+                return NotFound("Empty repository.");
             }
         }
 
         [HttpGet("{id}")]
-        public IActionResult Get(int id)
+        public IActionResult GetByEntityId(int id)
         {
             try
             {
@@ -44,7 +51,7 @@ namespace GeekMDSuite.WebAPI.Presentation.Controllers
             }
             catch (RepositoryElementNotFoundException)
             {
-                return NotFound();
+                return NotFound($"Cannot locate element with id {id}.");
             }
         }
 
@@ -57,17 +64,18 @@ namespace GeekMDSuite.WebAPI.Presentation.Controllers
                 UnitOfWork.Complete();
                 return Ok();
             }
-            catch (EntityNotUniqeException)
+            catch (EntityNotUniqeException e)
             {
-                return Conflict();
+                return Conflict(e.Message);
             }
             catch (ArgumentNullException)
             {
-                return BadRequest();
+                return BadRequest("A null entity was provided.");
             }
         }
 
         [HttpPut]
+        [Route("update/")]
         public IActionResult Put([FromBody] T entity)
         {
             try
@@ -76,17 +84,18 @@ namespace GeekMDSuite.WebAPI.Presentation.Controllers
                 UnitOfWork.Complete();
                 return Ok();
             }
-            catch (RepositoryElementNotFoundException)
+            catch (RepositoryElementNotFoundException e )
             {
-                return NotFound();
+                return NotFound(e.Message);
             }
             catch (ArgumentNullException)
             {
-                return BadRequest();
+                return BadRequest("A null entity was provided.");
             }
         }
 
-        [HttpDelete("{id}")]
+        [HttpDelete]
+        [Route("delete/{id}")]
         public IActionResult Delete(int id)
         {
             try
@@ -95,30 +104,34 @@ namespace GeekMDSuite.WebAPI.Presentation.Controllers
                 UnitOfWork.Complete();
                 return Ok();
             }
-            catch (RepositoryElementNotFoundException)
+            catch (RepositoryElementNotFoundException e)
             {
-                return NotFound();
+                return NotFound(e.Message);
             }
         }
 
         [HttpDelete]
+        //TODO: Remove this method
         public IActionResult Delete([FromBody] int[] ids)
         {
+            if (ids == null || ids.Length <= 0)
+                return new BadRequestResult();
+            
             try
             {
                 foreach (var id in ids) _repo.Delete(id);
                 UnitOfWork.Complete();
                 return Ok();
             }
-            catch (RepositoryElementNotFoundException)
+            catch (RepositoryElementNotFoundException e)
             {
-                return NotFound();
+                return NotFound(e.Message);
             }
         }
-
-        public ConflictResult Conflict()
+        
+        public ConflictResult Conflict(string message)
         {
-            return new ConflictResult();
+            return new ConflictResult(message);
         }
     }
 }
